Refactor token manager to support OAuth login use case

+++ src/edu/lternet/pasta/token/TokenManager.java
@@ -36,6 +36,7 @@ import org.apache.log4j.Logger;
 /**
  * @author servilla
  * @since Mar 9, 2012
+ * @modified Sept 19, 2019
  *        <p/>
  *        The TokenManager manages the NIS Data Portal "tokenstore" for
  *        storing user
@@ -51,28 +52,48 @@ public class TokenManager {
                                                               .token
                                                               .TokenManager
                                                               .class);
-    
+    private static String dbDriver;   // database driver
+    private static String dbURL;      // database URL
+    private static String dbUser;     // database user name
+    private static String dbPassword; // database user password
+
     /*
      * Instance variables
      */
 
-    private String dbDriver;   // database driver
-    private String dbURL;      // database URL
-    private String dbUser;     // database user name
-    private String dbPassword; // database user password
+    private String extToken = null;
+    private String b64Token = null;
+    private String token = null;
+    private String signature = null;
+    private String uid = null;
+    private String authSystem = null;
+    private Long ttl;
+    private ArrayList<String> groups;
 
     /*
      * Constructors
      */
 
-    public TokenManager() {
+    public TokenManager(String extToken) {
 
         Configuration options = ConfigurationListener.getOptions();
 
-        this.dbDriver = options.getString("db.Driver");
-        this.dbURL = options.getString("db.URL");
-        this.dbUser = options.getString("db.User");
-        this.dbPassword = options.getString("db.Password");
+        TokenManager.dbDriver = options.getString("db.Driver");
+        TokenManager.dbURL = options.getString("db.URL");
+        TokenManager.dbUser = options.getString("db.User");
+        TokenManager.dbPassword = options.getString("db.Password");
+
+        if (extToken != null) {
+            this.extToken = extToken;
+            this.b64Token = extToken.split("-")[0];
+            this.token = new String(Base64.decodeBase64(this.b64Token));
+            this.signature = extToken.split("-")[1];
+            String[] tokenParts = this.token.split("\\*");
+            this.uid = tokenParts[0];
+            this.authSystem = tokenParts[1];
+            this.ttl = new Long(tokenParts[2]);
+            this.groups = this.getGroups(token);
+        }
 
     }
     
@@ -80,47 +101,36 @@ public class TokenManager {
      * Methods
      */
 
-    /**
-     * @param args
-     */
-    public static void main(String[] args) {
+    public String getExtToken() {
+        return this.extToken;
+    }
 
-        ConfigurationListener.configure();
+    public String getB64Token() {
+        return this.b64Token;
+    }
 
-        String token = "sz46tDcFxqLby2TtlBARREdqGFSSRFbjSHPvMw0hgXLsG2uGlDW" +
-                       "rOzjf/zM7Yd7g4n8pK5qKzohvP9UdYqf/xyx/RBAUU1QYwmUXTA" +
-                       "5NnUZ5qHjYCtx3Y+DgwyNsQPoz6dQqR92BWWsWb39BilwfaRoyg" +
-                       "8vRbmJ3CFRslvB5WfUqEI2OIhD2h3VyYXq8V7f8X4IZDSHWMWNX" +
-                       "YMuxC3eQ+A==";
+    public String getToken() {
+        return this.token;
+    }
 
-        String username = "uid=ucarroll,o=LTER,dc=ecoinformatics,dc=org";
+    public String getSignature() {
+        return this.signature;
+    }
 
-        TokenManager tokenManager = new TokenManager();
+    public String getUid() {
+        return this.uid;
+    }
 
-        try {
-            tokenManager.setToken(username, token);
-        }
-        catch (SQLException e) {
-            logger.error(e);
-            e.printStackTrace();
-        }
-        catch (ClassNotFoundException e) {
-            logger.error(e);
-            e.printStackTrace();
-        }
+    public String getAuthSystem() {
+        return this.authSystem;
+    }
 
-        try {
-            System.out.println(tokenManager.getToken(username));
-        }
-        catch (SQLException e) {
-            logger.error(e);
-            e.printStackTrace();
-        }
-        catch (ClassNotFoundException e) {
-            logger.error(e);
-            e.printStackTrace();
-        }
+    public Long getTtl() {
+        return this.ttl;
+    }
 
+    public ArrayList<String> getGroups() {
+        return this.groups;
     }
 
     /**
@@ -128,7 +138,7 @@ public class TokenManager {
      *
      * @return The database Connection object.
      */
-    private Connection getConnection() throws ClassNotFoundException {
+    private static Connection getConnection() throws ClassNotFoundException {
 
         Connection conn = null;
 
@@ -136,7 +146,7 @@ public class TokenManager {
 
         // Load the jdbc driver.
         try {
-            Class.forName(this.dbDriver);
+            Class.forName(TokenManager.dbDriver);
         }
         catch (ClassNotFoundException e) {
             logger.error("Can't load driver " + e.getMessage());
@@ -145,8 +155,8 @@ public class TokenManager {
 
         // Make the database connection
         try {
-            conn = DriverManager.getConnection(this.dbURL, this.dbUser,
-                                                  this.dbPassword);
+            conn = DriverManager.getConnection(TokenManager.dbURL, TokenManager.dbUser,
+                    TokenManager.dbPassword);
 
             // If a SQLWarning object is available, print its warning(s).
             // There may be multiple warnings chained.
@@ -172,17 +182,14 @@ public class TokenManager {
     /**
      * Sets the token and user id into the tokenstore.
      *
-     * @param uid   The user identifier.
-     * @param token The base64 encrypted token.
      * @throws SQLException
      * @throws ClassNotFoundException
      */
-    public void setToken(String uid, String token) throws SQLException,
+    public void storeToken() throws SQLException,
                                                               ClassNotFoundException {
 
-        String sql = "SELECT authtoken.tokenstore.token FROM " +
-                         "authtoken.tokenstore WHERE " +
-                         "authtoken.tokenstore.user_id='" + uid + "';";
+        String sql = "SELECT authtoken.tokenstore.token FROM authtoken.tokenstore WHERE " +
+                         "authtoken.tokenstore.user_id='" + this.uid + "';";
 
         Connection dbConn = null; // database connection object
 
@@ -195,11 +202,8 @@ public class TokenManager {
                 if (rs.next()) {
 
                     // uid already in token store, perform "update".
-                    sql = "UPDATE authtoken.tokenstore " +
-                              "SET token='" + token + "'," +
-                              "date_created=now() " +
-                              "WHERE authtoken.tokenstore.user_id='" + uid +
-                              "';";
+                    sql = "UPDATE authtoken.tokenstore SET token='" + this.extToken + "',date_created=now() " +
+                              "WHERE authtoken.tokenstore.user_id='" + this.uid + "';";
 
                     if (stmt.executeUpdate(sql) == 0) {
                         SQLException e = new SQLException("setToken: update '"
@@ -211,13 +215,10 @@ public class TokenManager {
                 } else {
 
                     // uid not in token store, perform "insert".
-                    sql = "INSERT INTO authtoken.tokenstore VALUES " +
-                              "('" + uid + "','" + token + "', now());";
+                    sql = "INSERT INTO authtoken.tokenstore VALUES ('" + this.uid + "','" + this.extToken + "', now());";
 
                     if (stmt.executeUpdate(sql) == 0) {
-                        SQLException e = new SQLException("setToken: insert '"
-                                                              + sql + "' " +
-                                                              "failed");
+                        SQLException e = new SQLException("setToken: insert '" + sql + "' failed");
                         throw (e);
                     }
 
@@ -251,13 +252,12 @@ public class TokenManager {
      * @throws SQLException
      * @throws ClassNotFoundException
      */
-    public String getToken(String uid) throws SQLException,
+    public static String getExtToken(String uid) throws SQLException,
                                                   ClassNotFoundException {
 
         String token = null;
-        String sql = "SELECT authtoken.tokenstore.token FROM " +
-                         "authtoken.tokenstore WHERE " +
-                         "authtoken.tokenstore.user_id='" + uid + "';";
+        String sql = "SELECT authtoken.tokenstore.token FROM authtoken.tokenstore WHERE authtoken.tokenstore.user_id='"
+                + uid + "';";
 
         Connection dbConn = null; // database connection object
 
@@ -269,9 +269,7 @@ public class TokenManager {
                 if (rs.next()) {
                     token = rs.getString("token");
                 } else {
-                    SQLException e = new SQLException("getToken: uid '" + uid +
-                                                          "' not in authtoken" +
-                                                          ".tokenstore");
+                    SQLException e = new SQLException("getToken: uid '" + uid + "' not in authtoken.tokenstore");
                     throw (e);
                 }
             }
@@ -303,11 +301,10 @@ public class TokenManager {
      * @throws SQLException
      * @throws ClassNotFoundException
      */
-    public void deleteToken(String uid) throws SQLException,
+    public static void deleteToken(String uid) throws SQLException,
                                                    ClassNotFoundException {
 
-        String sql = "DELETE FROM authtoken.tokenstore WHERE " +
-                         "authtoken.tokenstore.user_id='" + uid + "';";
+        String sql = "DELETE FROM authtoken.tokenstore WHERE authtoken.tokenstore.user_id='" + uid + "';";
 
         Connection dbConn = null; // database connection object
 
@@ -316,8 +313,7 @@ public class TokenManager {
             try {
                 Statement stmt = dbConn.createStatement();
                 if (stmt.executeUpdate(sql) == 0) {
-                    SQLException e = new SQLException("deleteToken: delete '"
-                                                          + sql + "' failed");
+                    SQLException e = new SQLException("deleteToken: delete '" + sql + "' failed");
                     throw (e);
                 }
             }
@@ -340,179 +336,8 @@ public class TokenManager {
 
     }
 
-    /**
-     * Returns the clear text representation of the authentication token
-     * identified by the user
-     *
-     * @param uid The user identifier
-     * @return The clear text representation of the authentication token
-     * @throws SQLException
-     * @throws ClassNotFoundException
-     */
-    public String getCleartextToken(String uid) throws
-        SQLException, ClassNotFoundException {
-
-        String[] token = null;
-        String sql = "SELECT authtoken.tokenstore.token FROM " +
-                         "authtoken.tokenstore WHERE " +
-                         "authtoken.tokenstore.user_id='" + uid + "';";
-
-        Connection dbConn = null; // database connection object
-
-        try {
-            dbConn = getConnection();
-            try {
-                Statement stmt = dbConn.createStatement();
-                ResultSet rs = stmt.executeQuery(sql);
-                if (rs.next()) {
-                    token = rs.getString("token").split("-");
-                } else {
-                    SQLException e = new SQLException("getToken: uid '" + uid +
-                                                          "' not in authtoken" +
-                                                          ".tokenstore");
-                    throw (e);
-                }
-            }
-            catch (SQLException e) {
-                logger.error("getToken: " + e);
-                logger.error(sql);
-                e.printStackTrace();
-                throw (e);
-            }
-            finally {
-                dbConn.close();
-            }
-            // Will fail if database adapter class not found.
-        }
-        catch (ClassNotFoundException e) {
-            logger.error("getToken: " + e);
-            e.printStackTrace();
-            throw (e);
-        }
-
-        return new String(Base64.decodeBase64(token[0]));
-
-    }
-
-    /**
-     * Returns the base64 encoded digital signature of the authentication
-     * token identifier by the user
-     *
-     * @param uid The user identifier
-     * @return The base64 encoded digital signature
-     * @throws SQLException
-     * @throws ClassNotFoundException
-     */
-    public String getTokenSignature(String uid) throws
-        SQLException, ClassNotFoundException {
-
-        String[] token = null;
-        String sql = "SELECT authtoken.tokenstore.token FROM " +
-                         "authtoken.tokenstore WHERE " +
-                         "authtoken.tokenstore.user_id='" + uid + "';";
-
-        Connection dbConn = null; // database connection object
-
-        try {
-            dbConn = getConnection();
-            try {
-                Statement stmt = dbConn.createStatement();
-                ResultSet rs = stmt.executeQuery(sql);
-                if (rs.next()) {
-                    token = rs.getString("token").split("-");
-                } else {
-                    SQLException e = new SQLException("getToken: uid '" + uid +
-                                                          "' not in authtoken" +
-                                                          ".tokenstore");
-                    throw (e);
-                }
-            }
-            catch (SQLException e) {
-                logger.error("getToken: " + e);
-                logger.error(sql);
-                e.printStackTrace();
-                throw (e);
-            }
-            finally {
-                dbConn.close();
-            }
-            // Will fail if database adapter class not found.
-        }
-        catch (ClassNotFoundException e) {
-            logger.error("getToken: " + e);
-            e.printStackTrace();
-            throw (e);
-        }
-
-        return token[1];
-
-    }
-
-    /**
-     * Returns the user's LDAP distinguished name from the authentication token
-     *
-     * @param uid The user identifier
-     * @return The user's LDAP distinguished name
-     * @throws SQLException
-     * @throws ClassNotFoundException
-     */
-    public String getUserDistinguishedName(String uid) throws
-        SQLException, ClassNotFoundException {
-
-        String token = this.getCleartextToken(uid);
-        String[] tokenParts = token.split("\\*");
-
-        return tokenParts[0];
-    }
-
-    /**
-     * Returns the authentication system from the authentication token
-     *
-     * @param uid The user identifier
-     * @return The authentication system
-     * @throws SQLException
-     * @throws ClassNotFoundException
-     */
-    public String getTokenAuthenticationSystem(String uid) throws
-        SQLException, ClassNotFoundException {
-
-        String token = this.getCleartextToken(uid);
-        String[] tokenParts = token.split("\\*");
-
-        return tokenParts[1];
-    }
-
-    /**
-     * Returns the time-to-live (TTL) from the authentication token
-     *
-     * @param uid The user identifier
-     * @return The time-to-live
-     * @throws SQLException
-     * @throws ClassNotFoundException
-     */
-    public Long getTokenTimeToLive(String uid) throws
-        SQLException, ClassNotFoundException {
-
-        String token = this.getCleartextToken(uid);
-        String[] tokenParts = token.split("\\*");
-
-        return Long.valueOf(tokenParts[2]);
-    }
-
-    /**
-     * Returns the list of groups from the authentication token
-     *
-     * @param uid The user identifier
-     * @return The list of groups
-     * @throws SQLException
-     * @throws ClassNotFoundException
-     */
-    public ArrayList<String> getUserGroups(String uid)  throws
-        SQLException, ClassNotFoundException {
-
+    private ArrayList<String> getGroups(String token) {
         ArrayList<String> groups = new ArrayList<String>();
-
-        String token = this.getCleartextToken(uid);
         String[] tokenParts = token.split("\\*");
 
         for (int i = 3; i < tokenParts.length; i++) {
